Few more unit test refactored
